Some minor optimizations.

+++ src/move.h
@@ -25,7 +25,7 @@ typedef struct
 void move_to_string(move_t*, char*);
 void move_generate_moves(state_t*, move_t*, int*);
 uint64_t move_piece_moves(state_t*, int, int, int);
-int move_is_attacked(state_t*, uint64_t, int);
+int move_is_attacked(state_t*, int, int);
 
 void move_make(state_t*, move_t*);
 void move_unmake(state_t*, move_t*);
